Couple of bugs solved
The "models" variable has been changed to "persistent" and we do not use
"modelX": it does not work as a pointer, as we want.

Problem: we cannot use parfor with a persistent variable, so without it
the process time is doubled (up to 2s).

Still we need to fix the ghost effect