Added particle filter object, tests of initialisation
Particle filter object currently allows the initialisation of a filter using
the models, a number of particles, a bounding box and its centroid. Not sure if this
is exactly what is needed.

Basic tests on the object using the pf_tests script. Currently only looking at a single
bounding box, but the initialisation works correctly at least.