Multiple objects KF tracking
Complete and (hopefully) bug-free code to track multiple objects from
measurements. It creates a new KF for every new measurement that is not
assigned to a previous KF. It should delete KF's if they are not updated
after N_max frames (not yet tested).
Tested: track 2 objects entering the scene.