Particle filter deals with multiple measurements, but not completely tested
Particles are divided between the resampling step and being placed on measurements.

At the start of each timestep, particles are placed at locations corresponding to measurements received in the previous timestep. These particles are propagated in the same way as other particles taken from the resampling step.

Velocity no longer computed since we cannot yet match measurements in the previous timestep to those in the current one.