More cases in PF to allow simpler usage in frame processing loop
PF has an initialise flag which indicates when the particle filter has been initialised with some measurements. This is set in the constructor if centroids and bboxes are passed at that stage. Otherwise, it is set when the filter receives measurements to the step function. Currently, the flag is unset when the filter receives no measurements, but this will cause problems if you need to deal with multiple occluded objects.